Refactor k-progress
